# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "API de acesso aos dados do transporte público do estado do Espírito Santo",
		"version": "1.11.4",
		"title": "Nova api-ceturb"
	},
	"basePath": "/",
...
	"schemes": ["http"],
	"securityDefinitions": {},
	"paths": {
...
		"/agencias/contatos": {
			"get": {
				"summary": "Contatos",
				"description": "Lista os contatos que recebem logs do GTFS. \nOrigem: banco de dados",
				"responses": {
					"302": {
						"description": "Contatos encontrados",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Contato"
							}
						}
					},
					"404": {
						"description": "Contatos não encontrados",
						"schema": {
							"$ref": "#/definitions/ErrorMessage"
						}
					}
				},
				"tags": ["Agencias"],
				"produces": ["application/json"],
				"consumes": ["application/json"]
			}
		},
		"/agencias/feriados": {
			"get": {
				"summary": "Feriados",
				"description": "Lista os feriados previstos. \nOrigem: banco de dados",
				"responses": {
					"302": {
						"description": "Feriados encontrados",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Feriado"
							}
						}
					},
					"404": {
						"description": "Feriados não encontrados",
						"schema": {
							"$ref": "#/definitions/ErrorMessage"
						}
					}
				},
				"tags": ["Agencias"],
				"produces": ["application/json"],
				"consumes": ["application/json"]
			}
		},
		"/agencias/feriados/validar/{ano}/{mes}/{dia}": {
			"get": {
				"summary": "Consulta booleana de Feriados da agencia por data",
				"description": "Indica se a data consultada é um feriado registrado no banco auxiliar ou não.",
				"parameters": [{
					"name": "dia",
					"required": true,
					"in": "path",
					"description": "porção da data referente ao dia. exemplo: 25",
					"type": "number"
				}, {
					"name": "mes",
					"required": true,
					"in": "path",
					"description": "porção da data referente ao mês. exemplo: 12",
					"type": "number"
				}, {
					"name": "ano",
					"required": true,
					"in": "path",
					"description": "porção da data referente ao ano. exemplo: 2018",
					"type": "number"
				}],
				"responses": {
					"200": {
						"description": "Consulta executada com sucesso.",
						"schema": {
							"$ref": "#/definitions/FeriadoResponse"
						}
					}
				},
				"tags": ["Agencias"],
				"produces": ["application/json"],
				"consumes": ["application/json"]
			}
		},
		"/agencias/tarifas": {
			"get": {
				"summary": "Tarifas",
				"description": "Lista as tarifas. \nOrigem: banco de dados ",
				"responses": {
					"302": {
						"description": "Tarifas encontradas",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Tarifa"
							}
						}
					},
					"404": {
						"description": "Tarifas não encontradas",
						"schema": {
							"$ref": "#/definitions/ErrorMessage"
						}
					}
				},
				"tags": ["Agencias"],
				"produces": ["application/json"],
				"consumes": ["application/json"]
			}
		},
...
	},
	"definitions": {
...
		"Date": {
			"type": "object",
			"properties": {}
		},
		"ErrorMessage": {
			"type": "object",
			"properties": {
				"mensagem": {
					"type": "string"
				},
				"data": {
					"$ref": "#/definitions/Date"
				},
				"rota": {
					"type": "string"
				},
				"statusCode": {
					"type": "number"
				}
			},
			"required": ["mensagem", "data", "rota", "statusCode"]
		},
...
		"Contato": {
			"type": "object",
			"properties": {
				"id": {
					"type": "number"
				},
				"dataregistro": {
					"$ref": "#/definitions/Date"
				},
				"atualizadoem": {
					"$ref": "#/definitions/Date"
				},
				"nome": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"ativo": {
					"type": "boolean"
				},
				"agencia_id": {
					"type": "number"
				}
			},
			"required": ["id", "dataregistro", "atualizadoem", "nome", "email", "ativo", "agencia_id"]
		},
		"Feriado": {
			"type": "object",
			"properties": {
				"id": {
					"type": "number"
				},
				"dataregistro": {
					"$ref": "#/definitions/Date"
				},
				"atualizadoem": {
					"$ref": "#/definitions/Date"
				},
				"nome": {
					"type": "string"
				},
				"data": {
					"$ref": "#/definitions/Date"
				},
				"dataupload": {
					"$ref": "#/definitions/Date"
				},
				"agencia_id": {
					"type": "number"
				}
			},
			"required": ["id", "dataregistro", "atualizadoem", "nome", "data", "dataupload", "agencia_id"]
		},
		"FeriadoResponse": {
			"type": "object",
			"properties": {
				"data": {
					"$ref": "#/definitions/Date"
				},
				"feriado": {
					"type": "boolean"
				}
			},
			"required": ["data", "feriado"]
		},
		"Tarifa": {
			"type": "object",
			"properties": {
				"id": {
					"type": "number"
				},
				"dataregistro": {
					"$ref": "#/definitions/Date"
				},
				"atualizadoem": {
					"$ref": "#/definitions/Date"
				},
				"preco": {
					"type": "object"
				},
				"dataupload": {
					"$ref": "#/definitions/Date"
				},
				"agencia_id": {
					"type": "number"
				}
			},
			"required": ["id", "dataregistro", "atualizadoem", "preco", "dataupload", "agencia_id"]
		},
...
	}
}